# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Rest-Api for library",
        "title": "BibliotecaAPI",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
        "/api/admin/autores": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new author",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authors"
                ],
                "summary": "Create author",
                "parameters": [
                    {
                        "description": "Create an author",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.NewAuthor"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/data.Author"
                        }
                    }
                }
            }
        },
...
        "/api/admin/fines": {
            "get": {
                "description": "Get fines for a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Multas"
                ],
                "summary": "Get user fines",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "idsocio",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Fine"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/fines/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Search fines by user name or email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Multas"
                ],
                "summary": "Search fines by user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search term (name or email)",
                        "name": "busqueda",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Fine"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/fines/to": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all pending fines for admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Multas"
                ],
                "summary": "Get pending fines",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Fine"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/autores": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all authors",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authors"
                ],
                "summary": "Get all authors",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Author"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/fines": {
            "get": {
                "description": "Get pending fines for a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Multas"
                ],
                "summary": "Get user pending fines",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "usuario_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Fine"
                            }
                        }
                    }
                }
            }
        },
        "/api/fines/{id}/pay": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mark a fine as paid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Multas"
                ],
                "summary": "Pay fine",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Fine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment confirmation",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/api/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticacion"
                ],
                "summary": "Login a los Usuarios",
                "parameters": [
                    {
                        "description": "Credenciales de login",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Credentials"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticacion"
                ],
                "summary": "Registro Para nuevos usuarios",
                "parameters": [
                    {
                        "description": "Register credentials",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.RegisterRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "data.Author": {
            "type": "object",
            "properties": {
                "idautor": {
                    "type": "integer"
                },
                "nacionalidad": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
...
        "data.Fine": {
            "type": "object",
            "properties": {
                "estado": {
                    "type": "string"
                },
                "fechamulta": {
                    "type": "string"
                },
                "idmulta": {
                    "type": "integer"
                },
                "idprestamo": {
                    "type": "integer"
                },
                "saldopagar": {
                    "type": "number"
                }
            }
        },
...
        "data.NewAuthor": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                }
            }
        },
...
        "main.Credentials": {
            "type": "object",
            "properties": {
                "contrasena": {
                    "type": "string"
                },
                "correo": {
                    "type": "string"
                }
            }
        },
        "main.RegisterRequest": {
            "type": "object",
            "properties": {
                "contrasena": {
                    "type": "string"
                },
                "correo": {
                    "type": "string"
                },
                "direccion": {
                    "type": "string"
                },
                "fecha_nacimiento": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "rol": {
                    "type": "string"
                },
                "telefono": {
                    "type": "string"
                },
                "tipo_socio": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by your JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}